+++ crates/aviutl2/src/generic/bridge.rs
@@ -251,7 +251,12 @@ pub unsafe fn uninitialize_plugin_c_unwind<T: GenericSingleton>() {
 #[macro_export]
 macro_rules! register_generic_plugin {
     ($struct:ident, $($key:ident = $value:expr),* $(,)?) => {
-        ::aviutl2::__internal_module! {
+        ::$crate::__internal_module! {
+            #[unsafe(no_mangle)]
+            unsafe extern "C" fn RequiredVersion() -> u32 {
+                $struct::MINIMUM_AVIUTL2_VERSION.into()
+            }
+
             #[unsafe(no_mangle)]
             unsafe extern "C" fn InitializeLogger(logger: *mut $crate::sys::logger2::LOG_HANDLE) {
                 $crate::comptime_if::comptime_if! {