+++ crates/aviutl2/src/module/bridge.rs
@@ -167,7 +167,12 @@ extern "C" fn unreachable_function(_: *mut aviutl2_sys::module2::SCRIPT_MODULE_P
 #[macro_export]
 macro_rules! register_script_module {
     ($struct:ident, $($key:ident = $value:expr),* $(,)?) => {
-        ::aviutl2::__internal_module! {
+        ::$crate::__internal_module! {
+            #[unsafe(no_mangle)]
+            unsafe extern "C" fn RequiredVersion() -> u32 {
+                $struct::MINIMUM_AVIUTL2_VERSION.into()
+            }
+
             #[unsafe(no_mangle)]
             unsafe extern "C" fn InitializeLogger(logger: *mut $crate::sys::logger2::LOG_HANDLE) {
                 $crate::comptime_if::comptime_if! {